# server/swagger.yaml
openapi: 3.0.3
info:
  title: DLT TalentHub API
  description: |
    Backend API for DLT TalentHub - connecting talents with job opportunities in the DLT Africa ecosystem.
    
    ## Overview
    This API provides endpoints for user authentication, talent management, job postings, recruitment, and administrative functions.
    
    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15 minutes per IP address.
    
    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "status": "error",
      "message": "Error description",
      "errors": ["Detailed error messages"]
    }
    ```
  version: 1.0.0
  contact:
    name: DLT TalentHub Team
    email: support@dlt-talenthub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.dlt-talenthub.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/talents/profile:
    get:
      summary: Get current user's talent profile
      description: Retrieve the talent profile of the currently authenticated user
      tags: [Talents]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Talent profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Talent'
        '404':
          description: Talent profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update talent profile
      description: Update the talent profile of the currently authenticated user
      tags: [Talents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TalentUpdateRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Talent'
        '401':
          description: Not authenticated or not a talent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/talents/cv:
    post:
      summary: Upload CV
      description: Upload a CV document for the current talent
      tags: [Talents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cv:
                  type: string
                  format: binary
                  description: CV document (PDF, DOC, DOCX)
      responses:
        '200':
          description: CV uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      cv:
                        type: string
                        description: URL to the uploaded CV
        '400':
          description: Invalid file format or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete CV
      description: Remove the current talent's CV
      tags: [Talents]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CV deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: CV deleted successfully
  /api/talents/search:
    get:
      summary: Search talents
      description: Search for talents with various filters
      tags: [Talents]
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query for name or skills
        - in: query
          name: skills
          schema:
            type: string
          description: Filter by skills (comma-separated)
        - in: query
          name: location
          schema:
            type: string
          description: Filter by location
        - in: query
          name: availability
          schema:
            type: string
            enum: [available, busy, unavailable]
          description: Filter by availability status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      talents:
                        type: array
                        items:
                          $ref: '#/components/schemas/Talent'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/talents/skills/{skillName}:
    get:
      summary: Find talents by skill
      description: Find all talents who have a specific skill
      tags: [Talents]
      parameters:
        - in: path
          name: skillName
          required: true
          schema:
            type: string
          description: Skill name to search for
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: Talents with the specified skill
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      talents:
                        type: array
                        items:
                          $ref: '#/components/schemas/Talent'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/talents/available:
    get:
      summary: Get available talents
      description: Get a list of talents who are currently available for work
      tags: [Talents]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of available talents
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      talents:
                        type: array
                        items:
                          $ref: '#/components/schemas/Talent'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/talents/{talentId}:
    get:
      summary: Get talent profile by ID
      description: Retrieve a talent's profile by their ID
      tags: [Talents]
      parameters:
        - in: path
          name: talentId
          required: true
          schema:
            type: string
          description: Talent ID
      responses:
        '200':
          description: Talent profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Talent'
        '404':
          description: Talent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: Error message
        errors:
          type: array
          items:
            type: string
          description: Detailed error messages
    Talent:
      type: object
      properties:
        _id:
          type: string
          description: Talent ID
        userId:
          type: string
          description: Associated user ID
        skills:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: Skill name
              level:
                type: string
                enum: [beginner, intermediate, advanced, expert]
                description: Skill proficiency level
              yearsOfExperience:
                type: number
                minimum: 0
                description: Years of experience with this skill
        experience:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                description: Job title
              company:
                type: string
                description: Company name
              startDate:
                type: string
                format: date
                description: Start date
              endDate:
                type: string
                format: date
                description: End date (null if current)
              description:
                type: string
                description: Job description
        education:
          type: array
          items:
            type: object
            properties:
              degree:
                type: string
                description: Degree name
              institution:
                type: string
                description: Institution name
              field:
                type: string
                description: Field of study
              graduationYear:
                type: integer
                description: Graduation year
        availability:
          type: object
          properties:
            status:
              type: string
              enum: [available, busy, unavailable]
              description: Availability status
            startDate:
              type: string
              format: date
              description: Available start date
            endDate:
              type: string
              format: date
              description: Available end date
            hoursPerWeek:
              type: number
              minimum: 0
              maximum: 168
              description: Hours available per week
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum expected salary
            max:
              type: number
              minimum: 0
              description: Maximum expected salary
            currency:
              type: string
              default: USD
              description: Salary currency
        cv:
          type: string
          description: URL to CV document
        profileViews:
          type: integer
          minimum: 0
          description: Number of profile views
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TalentUpdateRequest:
      type: object
      properties:
        skills:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                description: Skill name
              level:
                type: string
                enum: [beginner, intermediate, advanced, expert]
                description: Skill proficiency level
              yearsOfExperience:
                type: number
                minimum: 0
                description: Years of experience with this skill
        experience:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                description: Job title
              company:
                type: string
                description: Company name
              startDate:
                type: string
                format: date
                description: Start date
              endDate:
                type: string
                format: date
                description: End date (null if current)
              description:
                type: string
                description: Job description
        education:
          type: array
          items:
            type: object
            properties:
              degree:
                type: string
                description: Degree name
              institution:
                type: string
                description: Institution name
              field:
                type: string
                description: Field of study
              graduationYear:
                type: integer
                description: Graduation year
        availability:
          type: object
          properties:
            status:
              type: string
              enum: [available, busy, unavailable]
              description: Availability status
            startDate:
              type: string
              format: date
              description: Available start date
            endDate:
              type: string
              format: date
              description: Available end date
            hoursPerWeek:
              type: number
              minimum: 0
              maximum: 168
              description: Hours available per week
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum expected salary
            max:
              type: number
              minimum: 0
              description: Maximum expected salary
            currency:
              type: string
              default: USD
              description: Salary currency
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        pages:
          type: integer
          minimum: 0
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrev:
          type: boolean
          description: Whether there is a previous page
...